Manage liquidity connect wallet redirect + combine manage liquidity
